프론트 게시판별 페이지 패키지 정리 #127
