+++ devbox_client/src/pages/education/EduDetail.jsx
@@ -1,0 +1,24 @@
+import '../../assets/img/apple-icon.png';
+import '../../assets/img/favicon.ico';
+import '../../assets/css/bootstrap.min.css';
+import '../../assets/css/boxicon.min.css';
+import '../../assets/css/templatemo.css';
+import '../../assets/css/custom.css';
+
+import Header from '../../components/Header';
+import Footer from '../../components/Footer';
+import DetailBasic from './DetailBasic';
+
+const EduDetail = () => {
+
+
+    return (
+        <div className="EduMain">
+            <Header />
+            <DetailBasic />
+            <Footer />
+        </div>
+    );
+};
+
+export default EduDetail;
