+++ devbox_client/src/pages/project/ProDetail.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
-import Footer from "../components/Footer";
-import Header from "../components/Header";
+import Footer from "../../components/Footer";
+import Header from "../../components/Header";
 import ProjectDetail from "./ProjectDetail";
 
 const ProDetail = () => {
