+++ devbox_client/src/pages/project/ProjectMain.jsx
@@ -1,11 +1,11 @@
-import backend from '../assets/img/backend.png';
-import frontend from '../assets/img/frontend.png';
-import devops from '../assets/img/devops.png';
-import cloud from '../assets/img/cloud.png';
-import data from '../assets/img/data.webp';
-import mobile from '../assets/img/mobile.png';
+import backend from '../../assets/img/backend.png';
+import frontend from '../../assets/img/frontend.png';
+import devops from '../../assets/img/devops.png';
+import cloud from '../../assets/img/cloud.png';
+import data from '../../assets/img/data.webp';
+import mobile from '../../assets/img/mobile.png';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
-import Button from '../components/Button';
+import Button from '../../components/Button';
 import { useEffect, useState } from 'react';
 
 const ProjectMain = (props) => {
